+++ glide/samplers/__init__.py
@@ -1,5 +1,6 @@
 from glide.samplers.active import ActiveSampler
 from glide.samplers.cost_optimal_random import CostOptimalRandomSampler
 from glide.samplers.stratified import StratifiedSampler
+from glide.samplers.uniform import UniformSampler
 
-__all__ = ["StratifiedSampler", "ActiveSampler", "CostOptimalRandomSampler"]
+__all__ = ["StratifiedSampler", "ActiveSampler", "CostOptimalRandomSampler", "UniformSampler"]